float fix

+++ mask2former/maskfiner_oracle_teacher_model.py
@@ -605,7 +605,7 @@ class MaskFinerOracleTeacher(nn.Module):
             edge_mask = self.compute_edge_mask_with_ignores(targets_shifted, border_mask)
             disagreement = self.count_edges_per_patch_masked(edge_mask, patch_size=patch_size)
             disagreement_map.append(disagreement)
-        disagreement_map = torch.stack(disagreement_map)
+        disagreement_map = torch.stack(disagreement_map).float()
         disagreement_map = (disagreement_map - disagreement_map.mean()) / disagreement_map.var().sqrt()
         #print("Initial disagreement map shape: {}".format(disagreement_map_tensor.shape))
         return disagreement_map
@@ -674,7 +674,7 @@ class MaskFinerOracleTeacher(nn.Module):
 
             disagreement = edge_mask_patched.sum(dim=(1, 2))
             disagreement_map.append(disagreement)
-        disagreement_map = torch.stack(disagreement_map)
+        disagreement_map = torch.stack(disagreement_map).float()
         disagreement_map = (disagreement_map - disagreement_map.mean()) / disagreement_map.var().sqrt()
 
         #print("Subsequent disagreement map shape: {}".format(disagreement_map_tensor.shape))
